Update requirements to modern versions.
With the many deprecation warnings, there is a need to limit the
user to certain versions. The wildcard syntax on the trailing
version number will allow for bug fixes in those packages without
allowing the user to install package versions that break things.

A separate issue is needed to tackle the DeprecationWarning
warnings coming from the responsible packages.

Fix issue #117.